Move deck-related code into deck.h
There is much inline code that we could consider moving into a deck.cpp,
but this can come later.
